+++ packages/core/src/Layout.tsx
@@ -6,53 +6,42 @@ import Footer from 'src/components/common/Footer'
 import Navbar from 'src/components/common/Navbar'
 import Toast from 'src/components/common/Toast'
 import RegionalizationBar from 'src/components/regionalization/RegionalizationBar'
-import { useUI } from 'src/sdk/ui'
-import { useModal } from 'src/sdk/ui/modal/Provider'
+import { useUI } from 'src/sdk/ui/Provider'
 
 const CartSidebar = lazy(() => import('src/components/cart/CartSidebar'))
-const RegionalizationModal = lazy(
+const RegionModal = lazy(
   () => import('src/components/regionalization/RegionalizationModal')
 )
 
 function Layout({ children }: PropsWithChildren<unknown>) {
-  const { displayMinicart } = useUI()
-  const { isRegionalizationModalOpen, setIsRegionalizationModalOpen } =
-    useModal()
+  const { cart: displayCart, modal: displayModal } = useUI()
 
   return (
     <>
-      <div id="layout">
-        <Alert
-          icon="Bell"
-          link={{ text: 'Buy now', to: '/office' }}
-          dismissible
-        >
-          Get 10% off today:&nbsp;<span>NEW10</span>
-        </Alert>
+      <Alert icon="Bell" link={{ text: 'Buy now', to: '/office' }} dismissible>
+        Get 10% off today:&nbsp;<span>NEW10</span>
+      </Alert>
 
-        <Navbar />
+      <Navbar />
 
-        <Toast />
+      <Toast />
 
-        <main>
-          <RegionalizationBar classes="display-mobile" />
-          {children}
-        </main>
+      <main>
+        <RegionalizationBar classes="display-mobile" />
+        {children}
+      </main>
 
-        <Footer />
+      <Footer />
 
-        {displayMinicart && (
-          <Suspense fallback={null}>
-            <CartSidebar />
-          </Suspense>
-        )}
-      </div>
-      {isRegionalizationModalOpen && (
+      {displayCart && (
         <Suspense fallback={null}>
-          <RegionalizationModal
-            isOpen={isRegionalizationModalOpen}
-            onDismiss={() => setIsRegionalizationModalOpen(false)}
-          />
+          <CartSidebar />
+        </Suspense>
+      )}
+
+      {displayModal && (
+        <Suspense fallback={null}>
+          <RegionModal />
         </Suspense>
       )}
     </>