+++ packages/core/src/components/cart/CartSidebar/CartSidebar.tsx
@@ -7,8 +7,8 @@ import Icon from 'src/components/ui/Icon'
 import SlideOver from 'src/components/ui/SlideOver'
 import { useCart } from 'src/sdk/cart/useCart'
 import { useCheckoutButton } from 'src/sdk/cart/useCheckoutButton'
-import { useUI } from 'src/sdk/ui'
-import { useModal } from 'src/sdk/ui/modal/Provider'
+import { useUI } from 'src/sdk/ui/Provider'
+import { useFadeEffect } from 'src/sdk/ui/useFadeEffect'
 
 import CartItem from '../CartItem'
 import EmptyCart from '../EmptyCart'
@@ -17,8 +17,8 @@ import OrderSummary from '../OrderSummary'
 function CartSidebar() {
   const btnProps = useCheckoutButton()
   const cart = useCart()
-  const { displayMinicart, closeMinicart } = useUI()
-  const { onModalClose } = useModal()
+  const { cart: displayCart, closeCart } = useUI()
+  const { fade, fadeOut } = useFadeEffect()
 
   const { items, totalItems, isValidating, subTotal, total } = cart
 
@@ -26,11 +26,13 @@ function CartSidebar() {
 
   return (
     <SlideOver
-      isOpen={displayMinicart}
-      onDismiss={closeMinicart}
+      fade={fade}
+      isOpen={displayCart}
+      onDismiss={fadeOut}
       size="partial"
       direction="rightSide"
       className="cart-sidebar"
+      onTransitionEnd={() => fade === 'out' && closeCart()}
     >
       <header data-testid="cart-sidebar">
         <div className="cart-sidebar__title">
@@ -41,7 +43,7 @@ function CartSidebar() {
           data-testid="cart-sidebar-button-close"
           aria-label="Close Cart"
           icon={<Icon name="X" width={32} height={32} />}
-          onClick={onModalClose}
+          onClick={fadeOut}
         />
       </header>
       <Alert icon={<Icon name="Truck" width={24} height={24} />}>
@@ -49,7 +51,7 @@ function CartSidebar() {
       </Alert>
 
       {isEmpty ? (
-        <EmptyCart onDismiss={onModalClose} />
+        <EmptyCart onDismiss={fadeOut} />
       ) : (
         <>
           <List>