+++ packages/core/src/components/common/Toast/Toast.tsx
@@ -1,7 +1,7 @@
 import { lazy, Suspense, useEffect } from 'react'
 
 import { useCart } from 'src/sdk/cart/useCart'
-import { useUI } from 'src/sdk/ui'
+import { useUI } from 'src/sdk/ui/Provider'
 
 const UIToast = lazy(() => import('src/components/ui/Toast'))