+++ packages/core/src/components/search/Filter/Filter.tsx
@@ -4,7 +4,8 @@ import { gql } from '@vtex/graphql-utils'
 import Button, { ButtonIcon } from 'src/components/ui/Button'
 import Icon from 'src/components/ui/Icon'
 import SlideOver from 'src/components/ui/SlideOver'
-import { useModal } from 'src/sdk/ui/modal/Provider'
+import { useUI } from 'src/sdk/ui/Provider'
+import { useFadeEffect } from 'src/sdk/ui/useFadeEffect'
 import type { Filter_FacetsFragment } from '@generated/graphql'
 
 import Facets from './Facets'
@@ -12,15 +13,6 @@ import { useFilter } from './useFilter'
 
 interface Props {
   facets: Filter_FacetsFragment[]
-  /*
-   * Control whether the filter modal is open. (mobile only)
-   */
-  isOpen?: boolean
-  /**
-   * This function is called whenever the user hits "Escape", clicks outside
-   * the filter modal or clicks in close button. (mobile only)
-   */
-  onDismiss?: () => void
   /**
    * ID to find this component in testing tools (e.g.: cypress,
    * testing-library, and jest).
@@ -28,20 +20,87 @@ interface Props {
   testId?: string
 }
 
-function Filter({
-  facets: allFacets,
-  onDismiss,
-  isOpen = false,
-  testId = 'store-filter',
-}: Props) {
+function FilterSlider({
+  facets,
+  testId,
+  dispatch,
+  expanded,
+  selected,
+}: Props & ReturnType<typeof useFilter>) {
+  const { closeFilter } = useUI()
+  const { fade, fadeOut } = useFadeEffect()
+
   const {
     setFacets,
-    toggleFacet,
     state: { selectedFacets },
   } = useSearch()
 
-  const { onModalClose } = useModal()
-  const { facets, selected, expanded, dispatch } = useFilter(allFacets)
+  return (
+    <SlideOver
+      isOpen
+      fade={fade}
+      onDismiss={fadeOut}
+      size="partial"
+      direction="rightSide"
+      className="filter-modal__content"
+      onTransitionEnd={() => fade === 'out' && closeFilter()}
+    >
+      <div className="filter-modal__body">
+        <header className="filter-modal__header">
+          <h2 className="text__lead">Filters</h2>
+          <ButtonIcon
+            data-testid="filter-modal-button-close"
+            aria-label="Close Filters"
+            icon={<Icon name="X" width={32} height={32} />}
+            onClick={() => {
+              dispatch({
+                type: 'selectFacets',
+                payload: selectedFacets,
+              })
+
+              fadeOut()
+            }}
+          />
+        </header>
+        <Facets
+          facets={facets}
+          testId={`mobile-${testId}`}
+          indicesExpanded={expanded}
+          onFacetChange={(facet) =>
+            dispatch({ type: 'toggleFacet', payload: facet })
+          }
+          onAccordionChange={(index) =>
+            dispatch({ type: 'toggleExpanded', payload: index })
+          }
+        />
+      </div>
+      <footer className="filter-modal__footer">
+        <Button
+          variant="secondary"
+          onClick={() => dispatch({ type: 'selectFacets', payload: [] })}
+        >
+          Clear All
+        </Button>
+        <Button
+          variant="primary"
+          data-testid="filter-modal-button-apply"
+          onClick={() => {
+            setFacets(selected)
+            fadeOut()
+          }}
+        >
+          Apply
+        </Button>
+      </footer>
+    </SlideOver>
+  )
+}
+
+function Filter({ facets: allFacets, testId = 'store-filter' }: Props) {
+  const filter = useFilter(allFacets)
+  const { toggleFacet } = useSearch()
+  const { filter: displayFilter } = useUI()
+  const { facets, expanded, dispatch } = filter
 
   return (
     <>
@@ -57,61 +116,7 @@ function Filter({
         />
       </div>
 
-      <SlideOver
-        isOpen={isOpen}
-        onDismiss={onDismiss}
-        size="partial"
-        direction="rightSide"
-        className="filter-modal__content"
-      >
-        <div className="filter-modal__body">
-          <header className="filter-modal__header">
-            <h2 className="text__lead">Filters</h2>
-            <ButtonIcon
-              data-testid="filter-modal-button-close"
-              aria-label="Close Filters"
-              icon={<Icon name="X" width={32} height={32} />}
-              onClick={() => {
-                dispatch({
-                  type: 'selectFacets',
-                  payload: selectedFacets,
-                })
-
-                onModalClose()
-              }}
-            />
-          </header>
-          <Facets
-            facets={facets}
-            testId={`mobile-${testId}`}
-            indicesExpanded={expanded}
-            onFacetChange={(facet) =>
-              dispatch({ type: 'toggleFacet', payload: facet })
-            }
-            onAccordionChange={(index) =>
-              dispatch({ type: 'toggleExpanded', payload: index })
-            }
-          />
-        </div>
-        <footer className="filter-modal__footer">
-          <Button
-            variant="secondary"
-            onClick={() => dispatch({ type: 'selectFacets', payload: [] })}
-          >
-            Clear All
-          </Button>
-          <Button
-            variant="primary"
-            data-testid="filter-modal-button-apply"
-            onClick={() => {
-              setFacets(selected)
-              onModalClose()
-            }}
-          >
-            Apply
-          </Button>
-        </footer>
-      </SlideOver>
+      {displayFilter && <FilterSlider {...filter} testId={testId} />}
     </>
   )
 }