+++ packages/core/src/components/ui/SlideOver/SlideOver.tsx
@@ -1,8 +1,5 @@
 import { Modal as UIModal } from '@faststore/ui'
-import { useEffect } from 'react'
 import type { ReactNode, HTMLAttributes } from 'react'
-
-import { useModal } from 'src/sdk/ui/modal/Provider'
 
 type Direction = 'leftSide' | 'rightSide'
 type WidthSize = 'full' | 'partial'
@@ -12,6 +9,7 @@ interface SlideOverProps extends HTMLAttributes<HTMLDivElement> {
   direction: Direction
   size: WidthSize
   children: ReactNode
+  fade: 'in' | 'out'
   /**
    * This function is called whenever the user clicks outside
    * the modal content
@@ -19,32 +17,21 @@ interface SlideOverProps extends HTMLAttributes<HTMLDivElement> {
   onDismiss?: () => void
 }
 
-const SlideOver = ({
+function SlideOver({
   isOpen,
-  onDismiss,
   direction = 'leftSide',
   size = 'full',
+  fade = 'out',
   children,
   ...otherProps
-}: SlideOverProps) => {
-  const { fade, onModalOpen, onModalClose } = useModal()
-
-  useEffect(() => {
-    isOpen && onModalOpen()
-  }, [isOpen, onModalOpen])
-
+}: SlideOverProps) {
   return (
     <UIModal
       isOpen={isOpen}
-      onDismiss={(e) => {
-        e.preventDefault()
-        onModalClose()
-      }}
       data-slide-over
       data-slide-over-direction={direction}
       data-slide-over-size={size}
       data-slide-over-state={fade}
-      onTransitionEnd={() => fade === 'out' && onDismiss?.()}
       {...otherProps}
     >
       {children}