+++ packages/core/src/sdk/ui/modal/index.ts
@@ -1,1 +1,0 @@
-export { default as ModalProvider } from './Provider'